% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_publications.R
\name{get_snsf_oa_work}
\alias{get_snsf_oa_work}
\title{Get SNSF OA monitoring data from Github}
\usage{
get_snsf_oa_work(
  repo,
  file,
  url = "https://raw.githubusercontent.com/snsf-data/",
  vars = c("doi", "id", "publication_year", "type")
)
}
\arguments{
\item{repo}{The name of the OA monitoring repository.}

\item{file}{The name of the data file.}

\item{url}{The url where is stored the raw content of the SNSF Github
page (default to "https://raw.githubusercontent.com/snsf-data/").}

\item{vars}{A vector with the name of the variable to keep from the SNSF Data
Portal.}
}
\value{
A data frame.
}
\description{
Get SNSF OA monitoring data from Github
}
