% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_publications.R
\name{get_snsf_work_dp}
\alias{get_snsf_work_dp}
\alias{get_snsf_work_dim}
\alias{get_snsf_work_oax}
\alias{get_snsf_work_cr}
\title{Get SNSF related scientific work from databases}
\usage{
get_snsf_work_dp(period)

get_snsf_work_dim(period, funder_id = "grid.425888.b", dim_token = NULL)

get_snsf_work_oax(period, funder_id = "F4320320924")

get_snsf_work_cr(period, funder_id = "10.13039/501100001711")
}
\arguments{
\item{period}{A numeric vector (can be a scalar) indicating the period in
years that must be search for scientific works (the period is defined based
on the minimum and maximum values in the vector).}

\item{funder_id}{A string with the ID of the funder. Each academic database
is expecting its own format:
\itemize{
\item Dimensions expects grid ID (default to SNSF grid ID: "grid.425888.b")
\item OpenAlex expects custom ID (default to SNSF funder ID: "F4320320924")
\item Crossref expects funder DOI (default to SNSF
DOI: "10.13039/501100001711").
}}

\item{dim_token}{The token to connect to the Dimensions API.}
}
\description{
Get SNSF related scientific work from databases
}
