% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensions.R
\name{pre_dim_work_query}
\alias{pre_dim_work_query}
\title{Function to get the Dimensions query with some pagination parameters
Format a Dimensions query for publications}
\usage{
pre_dim_work_query(period, funder_id = "grid.425888.b", skip = 0, limit = 1000)
}
\arguments{
\item{period}{A numeric vector (can be a scalar) indicating the period in
years that must be search for scientific works (the period is defined based
on the minimum and maximum values in the vector).}

\item{funder_id}{A string with the ID of the funder. Each academic database
is expecting its own format:
\itemize{
\item Dimensions expects grid ID (default to SNSF grid ID: "grid.425888.b")
\item OpenAlex expects custom ID (default to SNSF funder ID: "F4320320924")
\item Crossref expects funder DOI (default to SNSF
DOI: "10.13039/501100001711").
}}

\item{skip}{Integer with the number of publications to skip (useful for
pagination).}

\item{limit}{Integer with the number of publications to return (useful for
pagination).}
}
\description{
Function to get the Dimensions query with some pagination parameters
Format a Dimensions query for publications
}
\keyword{internal}
