% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensions.R
\name{query_dim_work}
\alias{query_dim_work}
\title{Function to query Dimension API with DSL query}
\usage{
query_dim_work(query, token)
}
\description{
Function to query Dimension API with DSL query
}
\keyword{internal}
