% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snsf.oa.ror.zurich.26-package.R
\docType{package}
\name{snsf.oa.ror.zurich.26-package}
\alias{snsf.oa.ror.zurich.26}
\alias{snsf.oa.ror.zurich.26-package}
\title{snsf.oa.ror.zurich.26: Code and data for an analysis of the SNSF OA monitoring presented at ROR Zürich 2026}
\description{
This package contains the code and data to reproduce an analysis presented at the Research on Research Event in Zürich on 2026/01/29. The analysis focuses on the coverage, within different academic databases, of scientific publications identified from Swiss National Science Foundation (SNSF) grants. Identifying these publications is critical for monitoring the open access of publications from SNSF grants.
}
\author{
\strong{Maintainer}: Simon Gorin \email{simon.gorin@snf.ch} (\href{https://orcid.org/0000-0001-7462-0446}{ORCID})

}
\keyword{internal}
